backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
